rename components folder to app
