+++ app/pages/index.tsx
@@ -1,4 +1,4 @@
-import Main from '../src/components/Main'
+import Main from '../src/app/Main'
 import Layout from '../src/layout/Layout'
 
 export default function Home() {
