+++ app/src/app/LensList/Table/Row/Aperture.tsx
