+++ app/src/app/LensList/Table/Row/Name.tsx
