+++ app/src/app/LensList/Table/Row/Row.tsx
