+++ app/src/app/LensList/Table/Row/Sensor.tsx
